fix(core): implement deleteCurrentNode command & fix node joining on Delete key (#3192)
* fix(core): :bug: fix delete on empty nodes joining next node incorrectly

This PR implements a "deleteCurrentNode" action in combination with registering this command inside the keymap for the delete key. This way, we editor will check, if the current node is empty before joining - if the current node is empty, the node will be removed. Joining will still work if the current node is not empty and the selection is at the end of the current node

2924

* refactor(core): :recycle: remove unnecessary return on handleDelete keymap